+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 - ♿️(frontend) restore skip to content link after header redesign #2510
 - 🌐(i18n) rename cn_CN to zh_CN, add eo_PL and zh_TW locales #2486
 - ✨(backend) conditional email notification in server to server api #2554
+- ✨(backend) profile api using django-silk
 
 ### Changed